requirements.txt
absl-py==0.8.1
astor==0.8.0
backcall==0.1.0
certifi==2019.9.11
cffi==1.13.0
cycler==0.10.0
decorator==4.4.1
et-xmlfile==1.0.1
gast==0.3.2
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.15.1
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.17.2
olefile==0.46
openpyxl==3.0.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
pyzmq==18.1.0
six==1.12.0
tensorboard==1.14.0
tensorboardX==1.9
tensorflow-estimator==1.14.0
tensorflow-gpu==0.12.1
termcolor==1.1.0
torch==0.4.1.post2
torchvision==0.2.1
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
